Add Background Music and Effect in main form and Add static Class name is "SoundEffect.cs" in folder Class/Sounds
