+++ BomberMan/BomberMan.csproj
@@ -33,6 +33,13 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="AxInterop.WMPLib, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\WMPLib.1.0.0\lib\AxInterop.WMPLib.dll</HintPath>
+    </Reference>
+    <Reference Include="Interop.WMPLib, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\WMPLib.1.0.0\lib\Interop.WMPLib.dll</HintPath>
+      <EmbedInteropTypes>True</EmbedInteropTypes>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -50,25 +57,26 @@
     <Compile Include="Class\Characters\Enemys.cs" />
     <Compile Include="Class\Characters\Player.cs" />
     <Compile Include="Class\Items\Items.cs" />
-    <Compile Include="Forms\Form1.cs">
+    <Compile Include="Class\Sounds\SoundEffect.cs" />
+    <Compile Include="Forms\MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Forms\Form1.Designer.cs">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <Compile Include="Forms\MainForm.Designer.cs">
+      <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Forms\Game.cs">
+    <Compile Include="Forms\GameForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Forms\Game.Designer.cs">
-      <DependentUpon>Game.cs</DependentUpon>
+    <Compile Include="Forms\GameForm.Designer.cs">
+      <DependentUpon>GameForm.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <EmbeddedResource Include="Forms\Form1.resx">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <EmbeddedResource Include="Forms\MainForm.resx">
+      <DependentUpon>MainForm.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Forms\Game.resx">
-      <DependentUpon>Game.cs</DependentUpon>
+    <EmbeddedResource Include="Forms\GameForm.resx">
+      <DependentUpon>GameForm.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
@@ -78,7 +86,9 @@
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
+      <DesignTime>True</DesignTime>
     </Compile>
+    <None Include="packages.config" />
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
       <LastGenOutput>Settings.Designer.cs</LastGenOutput>
@@ -94,6 +104,33 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Class\Maps\" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="resource\MusicEffect\Bgm.wav" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="resource\MusicEffect\Bgm2.wav" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="resource\MusicEffect\btnClick.wav" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="resource\MusicEffect\Click.wav" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="resource\MusicEffect\GameEnter.wav" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="resource\MusicEffect\Pass.wav" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="resource\MusicEffect\Result.wav" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="resource\MusicEffect\SongLogin.wav" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="resource\MusicEffect\Success.wav" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
