+++ BomberMan/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="WMPLib" version="1.0.0" targetFramework="net472" developmentDependency="true" />
+</packages>
