+++ BomberMan/resource/FileData/Highscore.txt
@@ -1,0 +1,1 @@
+sdaw/wdwad/prem/1234/0/0/0/0&
